build: ml: bug 26194: mobs support in blasr makefiles
This adds support for mobs to control the blasr makefile from the
make command line and adds support for building blasr using shared
libraries for its dependencies.

This basically just copies local copy of common.mk that the mobs build
was using, so the mobs part of it will still work the same.

Please review this change before I check it in.  It should not affect
the current default operation of the makefile, it mostly just exposes
variables that can be controlled from the command line. 

NOTE: I need to make a similar change to pbdagcon.  But since that is
      now a github only project, perhaps we can discuss how to best
      make/review that change.

#review-154315
@cdunn
@yli
@jdrake