Blasr: avoid non-full-pass subreads as concordant template.
Blasr should use 'full-pass' subreads as concordant
template if possible. However this info is not available in
subreads.bam. Since we know that the first and last subreads
are less likely to be full-pass, here we changed blasr so that
it avoid using the first and last subreads as concordant
template when input is subreads.bam.

update cram tests for bam concordant.
update libcpp